requirement.txt
matplotlib==2.2.3
numpy==1.14.5
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
Pillow==5.2.0
scikit-learn==0.19.2
scipy==1.1.0
tensorflow-gpu==1.10.1
